refactor: simplified naming
