+++ cognite/neat/rules/models/dms/__init__.py
@@ -1,6 +1,6 @@
-from ._dms_architect_rules import DMSContainer, DMSMetadata, DMSProperty, DMSRules, DMSView
-from ._dms_rules_input import DMSContainerInput, DMSMetadataInput, DMSPropertyInput, DMSRulesInput, DMSViewInput
-from ._dms_schema import DMSSchema, PipelineSchema
+from ._rules import DMSContainer, DMSMetadata, DMSProperty, DMSRules, DMSView
+from ._rules_input import DMSContainerInput, DMSMetadataInput, DMSPropertyInput, DMSRulesInput, DMSViewInput
+from ._schema import DMSSchema, PipelineSchema
 
 __all__ = [
     "DMSRules",
