+++ cognite/neat/rules/models/information/__init__.py
@@ -1,3 +1,3 @@
-from ._information_rules import InformationClass, InformationMetadata, InformationProperty, InformationRules
+from ._rules import InformationClass, InformationMetadata, InformationProperty, InformationRules
 
 __all__ = ["InformationRules", "InformationMetadata", "InformationClass", "InformationProperty"]
